Renames `Image` to `ImageLoader`
