+++ components.d.ts
@@ -38,6 +38,7 @@ declare module 'vue' {
     'ILogos:bluesky': typeof import('~icons/logos/bluesky')['default']
     Image: typeof import('./src/components/Utilities/Image.vue')['default']
     ImageContainer: typeof import('./src/components/Utilities/ImageContainer.vue')['default']
+    ImageLoader: typeof import('./src/components/Utilities/ImageLoader.vue')['default']
     'IMdi:accounts': typeof import('~icons/mdi/accounts')['default']
     'IMdi:alertCircle': typeof import('~icons/mdi/alert-circle')['default']
     'IMdi:block': typeof import('~icons/mdi/block')['default']
