+++ src/components/Feed/CustomFeedButton.vue
@@ -10,7 +10,7 @@
         <div class="flex flex-col gap-2 w-full overflow-hidden">
             <div class="flex gap-1 items-center">
                 <div class="flex bg-blueskyBlue aspect-square shrink-0 w-8 items-center justify-center">
-                    <Image v-if="typeof feedGeneratorView != 'undefined' && typeof feedGeneratorView.avatar != 'undefined'"
+                    <ImageLoader v-if="typeof feedGeneratorView != 'undefined' && typeof feedGeneratorView.avatar != 'undefined'"
                     :img-url="feedGeneratorView.avatar" loader-type="spinner" :fill-container="true"/>
                     <i-mingcute:radar-2-fill v-else class="text-white h-6 w-6"/>
                 </div>
@@ -48,7 +48,7 @@
 import { defineComponent, PropType } from 'vue'
 import { AppBskyFeedDefs } from '@atproto/api/dist/client';
 import { getCompactNumberValue } from '../../helpers/converters';
-import Image from '../Utilities/Image.vue';
+import ImageLoader from '../Utilities/ImageLoader.vue';
 
 export default defineComponent({
     props:{
