+++ src/components/Utilities/EmbedExternal.vue
@@ -7,7 +7,7 @@
         border-outline hover:border-embedHoverBorder hover:bg-embedHoverBG bg-postBG
         overflow-hidden text-xs cursor-pointer">
             <div class="relative border-b-[1px] border-outline aspect-[1.91/1]">
-                <Image :img-url="typeof embed != 'undefined' && typeof embed.external != 'undefined' && typeof embed.external.thumb != 'undefined' ? embed.external.thumb : ''"
+                <ImageLoader :img-url="typeof embed != 'undefined' && typeof embed.external != 'undefined' && typeof embed.external.thumb != 'undefined' ? embed.external.thumb : ''"
                 class="absolute w-full h-full object-center object-cover" :fill-container="true"/>
             </div>
             <div class="p-2 font-normal">
@@ -29,7 +29,7 @@
         border-outline hover:border-embedHoverBorder hover:bg-embedHoverBG bg-postBG
         overflow-hidden text-xs cursor-pointer">
             <div class="relative border-b-[1px] border-outline aspect-[1.91/1]">
-                <Image :img-url="typeof embed != 'undefined' && typeof embed.external != 'undefined' && typeof embed.external.thumb != 'undefined' ? embed.external.thumb : ''"
+                <ImageLoader :img-url="typeof embed != 'undefined' && typeof embed.external != 'undefined' && typeof embed.external.thumb != 'undefined' ? embed.external.thumb : ''"
                 class="absolute w-full h-full object-center object-cover" :fill-container="true"/>
             </div>
             <div class="p-2 font-normal">
@@ -74,7 +74,7 @@ import { OptionsMenuState } from '../../state/OptionsMenuState.vue';
 import { IOptionMenuItem, ItemType } from './OptionsMenu.vue';
 import { CopyTextToClipboard } from '../../state/AppState.vue';
 import { openUrl } from '@tauri-apps/plugin-opener';
-import Image from './Image.vue';
+import ImageLoader from './ImageLoader.vue';
 
 /**
  * Method used to open link in the system's default browser.
